Added map service, openweathermap api call working as well
