+++ app/controllers/api/v1/weather_controller.rb
@@ -1,7 +1,5 @@
 class Api::V1::WeatherController < ApplicationController
   def get_forecast
     WeatherFacade.new(params[:location])
-    # coordinates = WeatherService.lat_and_long(params[:location])
-    # require 'pry'; binding.pry
   end
 end
