+++ app/facades/weather_facade.rb
@@ -5,10 +5,15 @@ class WeatherFacade
   def initialize(location)
     @location = location
     @lat_and_long = get_location
+    @forecast = get_forecast
   end
 
   def get_location
-    attributes = WeatherService.lat_and_long(@location)
+    attributes = MapService.lat_and_long(@location)
     @lat_and_long = MapQuest.new(attributes)
+  end
+
+  def get_forecast
+    forecast = WeatherService.get_forecast(@lat_and_long)
   end
 end
